feat(hooks): detect prefixed git commit in shell payload scanner
Detect no-verify git commits when shell payloads prefix the command with wrappers like sudo, env, nice, ionice, strace, ltrace, time, command, or exec. Also align the ACP heuristic with the shell scanner, split lefthook parsing out of run_cmd_shell.rs to satisfy monolith gates, and bump the workspace patch version required by pre-commit.
